chore(wallet)_: move route execution code to separate module
